+++ goondex/pkg/crawler/interface.go
