+++ goondex/pkg/engine/engine_test.go
@@ -1,31 +1,40 @@
 package engine
 
 import (
+	"goondex/engine/storage"
+	"goondex/index"
+	"goondex/web"
 	"testing"
 )
 
-type StubScanner struct {
+type StubCrawler struct {
 }
 
-func (st *StubScanner) Scan(string, int) (map[string]string, error) {
-	links := map[string]string{
-		"yandex.ru":  "Яндекс",
-		"google.com": "Google",
+func (st *StubCrawler) Scan(string, int) ([]web.Page, error) {
+	links := []web.Page{
+		web.Page{ID: 1, URL: "yandex.ru", Title: "Яндекс"},
+		web.Page{ID: 2, URL: "google.com", Title: "Google"},
 	}
 	return links, nil
 }
 
-func Test_Search(t *testing.T) {
+func TestSearch(t *testing.T) {
 	eng := Engine{
-		scanner: &StubScanner{},
+		crawler: &StubCrawler{},
+		index:   index.New(),
+		storage: storage.New(),
 	}
 	eng.Scan("example.com")
 	want := "Яндекс"
-	got := eng.Search("yandex")
-	if len(got) != 1 {
-		t.Fatalf("длина хэша не равна 1")
+	result, err := eng.Search("Яндекс")
+	if err != nil {
+		t.Fatalf("поиск вернул ошибку")
 	}
-	if got["yandex.ru"] != want {
-		t.Fatalf("для ключа yandex.ru ждали %s, получили %s", want, got["yandex.ru"])
+	if len(result) != 1 {
+		t.Fatalf("длина результата не равна 1")
+	}
+	got := result[0].Title
+	if got != want {
+		t.Fatalf("для ключа yandex.ru ждали %s, получили %s", want, got)
 	}
 }
