+++ goondex/pkg/engine/go.mod
@@ -1,10 +1,3 @@
-module goondex/pkg/engine
+module engine
 
 go 1.15
-
-replace goondex/pkg/spider v0.0.0 => ../spider
-
-require (
-	golang.org/x/net v0.0.0-20201016165138-7b1cca2348c0 // indirect
-	goondex/pkg/spider v0.0.0
-)
