--- goondex/pkg/engine/scanner.go
@@ -1,6 +1,0 @@
-package engine
-
-// Позволяет сканировать сайты
-type Scanner interface {
-	Scan(string, int) (map[string]string, error)
-}
