+++ goondex/pkg/index/interface.go
