--- goondex/pkg/spider/go.mod
@@ -1,5 +1,0 @@
-module goondex/pkg/spider
-
-go 1.15
-
-require golang.org/x/net v0.0.0-20201016165138-7b1cca2348c0
