--- goondex/pkg/spider/spider_test.go
@@ -1,23 +1,0 @@
-// Package spider реализует сканер содержимого веб-сайтов.
-
-// Пакет позволяет получить список ссылок и заголовков страниц внутри веб-сайта по его URL.
-
-package spider
-
-import (
-	"testing"
-)
-
-func TestScanSite(t *testing.T) {
-	const url = "https://habr.com"
-	const depth = 2
-	spd := &Spider{}
-	data, err := spd.Scan(url, depth)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	for k, v := range data {
-		t.Logf("%s -> %s\n", k, v)
-	}
-}
