--- review/cmd/cli/main.go
@@ -1,73 +1,0 @@
-package main
-
-import (
-	"bufio"
-	"flag"
-	"fmt"
-	"log"
-	"os"
-
-	"goondex/engine"
-	"goondex/web"
-)
-
-func main() {
-	const depth = 2
-
-	var url string
-	flag.StringVar(&url, "url", "https://bash.im", "Адрес сайта для сканирования")
-	flag.Parse()
-
-	eng, err := initEngine(url)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	// Флаг, который определяет выполнять ли поиск или нет
-	runSearch := true
-	query := ""
-
-	for runSearch {
-		query, runSearch = userInput()
-
-		if runSearch {
-			fmt.Println("[Поиск] Запрос -", query)
-			res, err := eng.Search(query)
-			if err == nil {
-				printResult(res)
-			} else {
-				fmt.Printf("[Ошибка] Поиск вернул ошибку %v\n", err)
-			}
-		}
-	}
-}
-
-func initEngine(url string) (eng *engine.Engine, err error) {
-	eng = engine.New()
-	fmt.Println("[Сканирование] Страница -", url)
-	err = eng.Scan(url)
-	if err != nil {
-		return nil, fmt.Errorf("ошибка при сканировании сайта %s: %v", url, err)
-	}
-	return eng, nil
-}
-
-func printResult(result []web.Page) {
-	fmt.Println("[Результат]")
-	for index, page := range result {
-		fmt.Printf("%d. %s\n", index, page.URL)
-		fmt.Println(page.Title)
-	}
-}
-
-func userInput() (input string, runSearch bool) {
-	const exitCommand = "/q"
-
-	fmt.Println("[Ввод запроса] Введите запрос или /q для выхода:")
-	fmt.Print("> ")
-
-	scanner := bufio.NewScanner(os.Stdin)
-	scanner.Scan()
-	input = scanner.Text()
-	return input, input != exitCommand
-}
