--- review/pkg/engine/go.mod
@@ -1,3 +1,0 @@
-module engine
-
-go 1.15
